# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: AgentOps API
  version: 0.2.0
  description: |
    Multi-agent CI/CD orchestrator: review, test, approval, blue/green deploy,
    incident detection and AI-driven root cause analysis.

    All list endpoints support cursor-based pagination via `?cursor=` and `?limit=`.
servers:
  - url: http://localhost:4000
    description: Local dev
paths:
...
  /api/pipelines:
    get:
      summary: List pipelines (paginated, filterable)
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [running, complete, blocked, awaiting_approval, failed] } }
        - { name: repo, in: query, schema: { type: string } }
        - { name: risk_level, in: query, schema: { type: string, enum: [auto, soft, hard] } }
      responses:
        '200':
          description: Paginated pipeline list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /api/pipelines/{id}/decisions:
    get:
      summary: Get all agent decisions for a pipeline
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Agent decisions list
...
components:
  schemas:
    PaginatedResponse:
      type: object
      properties:
        items: { type: array, items: { type: object } }
        count: { type: integer }
        next_cursor: { type: string, nullable: true }
...